chore(`btcdctrl`): rename to `btcdtest`
