+++ btcdtest/btcdtest_test.go
@@ -1,4 +1,4 @@
-package btcdctrl
+package btcdtest
 
 import (
 	"os"
